requirements.txt
discord.py
openai
tmdbv3api
arrapi
regex
pyyaml
flask
gunicorn
imdbpy